Stripe payment working but order not done

+++ controllers/orderControllers.js
@@ -1,5 +1,7 @@
 const Order = require('../models/order');
 const Cart = require('../models/Cart');
+const User = require('../models/User');
+const stripe = require('stripe')('sk_test_bO1PBN9aUGxGlERZP69pn8ky00xB9CEove');
 
 module.exports.get_orders = async (req,res) => {
     const userId = req.params.id;
@@ -9,18 +11,27 @@ module.exports.get_orders = async (req,res) => {
 module.exports.checkout = async (req,res) => {
     try{
         const userId = req.params.id;
+        const source = req.body;
         let cart = await Cart.findOne({userId});
-
-        // no payments for now. Will add later!
-        // for now just move cart to order!
+        let user = await User.findOne({_id: userId});
+        const email = user.email;
         if(cart){
-            const order = await Order.create({
-                userId,
-                items: cart.items,
-                bill: cart.bill
-            });
-            const data = await Cart.findByIdAndDelete({_id:cart.id});
-            return res.status(201).send(order);
+            const charge = await stripe.charges.create({
+                amount: cart.bill,
+                currency: 'inr',
+                source,
+                receipt_email: email
+            })
+            if(!charge) throw Error('Payment failed');
+            if(charge){
+                const order = await Order.create({
+                    userId,
+                    items: cart.items,
+                    bill: cart.bill
+                });
+                const data = await Cart.findByIdAndDelete({_id:cart.id});
+                return res.status(201).send(order);
+            }
         }
         else{
             res.status(500).send("You do not have items in cart");
